+++ src/components/switcher/switcher-component.jsx
@@ -3,10 +3,10 @@ import Link from 'redux-first-router-link';
 import styles from './switcher-styles.module';
 
 const Switcher = ({ route, handleClick }) => {
-  const isDataGlobe = route.path === '/simple';
+  const isDataGlobe = route.path === '/dataGlobe';
   const experienceText = isDataGlobe ? 'featured maps': 'explore data';
   return (
-    <Link className={styles.switcherContainer} to={{ type: isDataGlobe ? 'location/FEATURED' : 'location/SIMPLE'}}>
+    <Link className={styles.switcherContainer} to={{ type: isDataGlobe ? 'location/FEATURED' : 'location/DATA'}}>
       <div className={styles.switcher}>
         Go To {experienceText}
       </div>
