+++ src/layouts/app-layout/app-layout-component.jsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import Proptypes from 'prop-types';
-import FeaturedGlobe from 'pages/featured-globe-simple';
-import DataGlobeSimple from 'pages/simple';
+import FeaturedGlobe from 'pages/featured-globe';
+import DataGlobe from 'pages/data-globe';
 import MapIframe from 'pages/map-iframe';
 
 class App extends PureComponent {
@@ -9,8 +9,8 @@ class App extends PureComponent {
     const { route } = this.props;
     const { page } = route;
     const embedded = page === 'map-iframe';
-    const whichGlobe = embedded ? <MapIframe /> : <DataGlobeSimple />;
-    return page === 'featured-globe-simple' ? <FeaturedGlobe /> : whichGlobe;
+    const whichGlobe = embedded ? <MapIframe /> : <DataGlobe />;
+    return page === 'featured-globe' ? <FeaturedGlobe /> : whichGlobe;
   }
 }
