+++ src/pages/data-globe/data-globe-component.jsx
@@ -23,7 +23,7 @@ const ProtectedAreasTooltips = loadable(() => import('components/protected-areas
 
 const { REACT_APP_ARGISJS_API_VERSION:API_VERSION } = process.env
 
-const DataGlobeComponentSimple = ({
+const DataGlobeComponent = ({
   sceneSettings,
   isFullscreenActive,
   isSidebarOpen,
@@ -91,4 +91,4 @@ const DataGlobeComponentSimple = ({
   )
 }
 
-export default DataGlobeComponentSimple;
+export default DataGlobeComponent;
