+++ src/pages/data-globe/data-globe-initial-state.js
