+++ src/pages/featured-globe/featured-globe-component.jsx
@@ -32,7 +32,7 @@ const ProtectedAreasTooltips = loadable(() => import('components/protected-areas
 
 const { REACT_APP_ARGISJS_API_VERSION:API_VERSION } = process.env
 
-const DataGlobeComponentSimple = ({
+const DataGlobeComponent = ({
   sceneSettings,
   isFullscreenActive,
   selectedFeaturedMap,
@@ -149,4 +149,4 @@ const DataGlobeComponentSimple = ({
   )
 }
 
-export default DataGlobeComponentSimple;
+export default DataGlobeComponent;
