+++ src/pages/featured-globe/featured-globe-initial-state.js
